[Bug 139] File aliases merge with HEAD
git-svn-id: svn+ssh://svn.openmicroscopy.org.uk/home/svn/ome/trunk@4779 61f9eca8-113b-463c-a1a4-bab9b0a29f14

+++ src/C/omeis/File.c
@@ -52,6 +52,14 @@
 
 
 #include "File.h"
+#include "OMEIS_Error.h"
+
+/* Private prototypes */
+int update_file_info (FileRep *myFile);
+OID check_aliases (OID ID, const char *name);
+int make_alias (OID ID,FileRep *aliasFile);
+char *get_rel_path (const char *toPath, const char *fromPath, char *pathBuf);
+
 
 void freeFileRep (FileRep *myFile)
 {
@@ -64,6 +72,8 @@ void freeFileRep (FileRep *myFile)
 	
 	if (myFile->fd_info >=0 ) close (myFile->fd_info);
 	if (myFile->fd_rep >=0 )  close (myFile->fd_rep);
+	
+	if (myFile->aliases) free (myFile->aliases);
 	free (myFile);
 }
 
@@ -93,7 +103,8 @@ char *sha1DBfile="Files/sha1DB.idx";
 	/* If we got an ID, set the paths */
 	if (ID) {
 		if (! getRepPath (ID,myFile->path_rep,0)) {
-			fprintf (stderr,"Could not get path to files file.\n");
+			OMEIS_DoError ("Could not get path for FIleID=%llu: %s",
+				(unsigned long long)ID,strerror (errno));
 			freeFileRep (myFile);
 			return (NULL);
 		}
@@ -114,13 +125,13 @@ struct stat fStat;
 	if (!ID) return (NULL);
 
 	if (! (myFile = newFileRep (ID))) {
-		fprintf (stderr,"Could not get a File object.\n");
+		OMEIS_DoError ("Could not get a File object (no memory).");
 		return (NULL);
 	}
 
 	if ( (myFile->fd_rep = openRepFile (myFile->path_rep, O_RDONLY)) < 0) {
 		freeFileRep (myFile);
-		fprintf (stderr,"Could not open repository file (FileID = %llu: %s\n", (unsigned long long)ID,strerror( errno ));
+		OMEIS_DoError ("Could not open repository file (FileID = %llu: %s", (unsigned long long)ID,strerror( errno ));
 		return (NULL);
 	}
 
@@ -128,7 +139,7 @@ struct stat fStat;
 	lockRepFile (myFile->fd_rep, 'r', offset, length);
 
 	if (fstat (myFile->fd_rep, &fStat) < 0) {
-		fprintf (stderr,"Could not get size of FileID=%llu",(unsigned long long)myFile->ID);
+		OMEIS_DoError ("Could not get size of FileID=%llu",(unsigned long long)myFile->ID);
 		freeFileRep (myFile);
 		return (NULL);			
 	}
@@ -138,7 +149,7 @@ struct stat fStat;
         length = myFile->size_rep;
 
     if (offset+length > myFile->size_rep) {
-        fprintf (stderr,"Trying to read past end of file\n");
+        OMEIS_DoError ("Trying to read past end of FileID=%llu",(unsigned long long)myFile->ID);
         freeFileRep (myFile);
         return (NULL);
     }
@@ -146,7 +157,7 @@ struct stat fStat;
     myFile->size_buf = myFile->size_rep;
 
 	if ( (myFile->file_buf = mmap (NULL, myFile->size_rep, PROT_READ, MAP_SHARED, myFile->fd_rep, 0LL)) == (void *) -1 ) {
-		fprintf (stderr,"Could not mmap FileID=%llu",(unsigned long long)myFile->ID);
+		OMEIS_DoError ("Could not mmap FileID=%llu",(unsigned long long)myFile->ID);
 		freeFileRep (myFile);
 		return (NULL);			
 	}
@@ -217,7 +228,9 @@ OID existOID;
 FileRep *NewFile (char *filename, size_t size)
 {
 FileRep *myFile;
+FileInfo *myInfo;
 char error[OMEIS_ERROR_SIZE];
+
 
 	if ( size > UINT_MAX || size == 0) return NULL;  /* Bad mojo for mmap */
 
@@ -227,11 +240,11 @@ char error[OMEIS_ERROR_SIZE];
 	}
 	myFile->ID = nextID(myFile->path_ID);
 	if (myFile->ID <= 0 && errno) {
-		perror ("Couldn't get next File ID");
+		OMEIS_DoError ("Couldn't get next File ID");
 		freeFileRep (myFile);
 		return (NULL);
 	} else if (myFile->ID <= 0){
-		fprintf (stderr,"Happy New Year !!!\n");
+		OMEIS_DoError ("Happy New Year !!!");
 		freeFileRep (myFile);
 		return (NULL);
 	}
@@ -260,41 +273,421 @@ char error[OMEIS_ERROR_SIZE];
 
 	if ( (myFile->file_buf = (char *)mmap (NULL, size, PROT_READ|PROT_WRITE , MAP_SHARED, myFile->fd_rep, 0LL)) == (char *) -1 ) {
 		DeleteFile (myFile);
-		fprintf (stderr,"Couldn't mmap File %s (ID=%llu)\n",myFile->path_rep,
+		OMEIS_DoError ("Couldn't mmap File %s (ID=%llu)",myFile->path_rep,
 			(unsigned long long)myFile->ID);
 		freeFileRep (myFile);
 		return (NULL);
 	}
-	
+
+	myInfo = &(myFile->file_info);
+	memset(myInfo, 0, sizeof(myInfo));
+
+	myInfo->mySig    = OME_IS_FILE_SIG;
+	myInfo->vers     = OME_IS_FILE_VER;
+	myInfo->ID       = myFile->ID;
 	if (filename)
-		strncpy (myFile->file_info.name,filename,OMEIS_PATH_SIZE-1);
-	else
-		strcpy (myFile->file_info.name,"");
+		strncpy (myInfo->name,filename,OMEIS_PATH_SIZE-1);
+	if ( write (myFile->fd_info,(void *)myInfo,sizeof(FileInfo)) != sizeof(FileInfo) ) {
+		DeleteFile (myFile);
+		OMEIS_DoError ("Couldn't write info for File %s (ID=%llu)",myFile->path_info,
+			(unsigned long long)myFile->ID);
+		freeFileRep (myFile);
+		return (NULL);
+	}
+	close (myFile->fd_info);
+	myFile->fd_info = -1;
 
 	return (myFile);
 }
 
+
+
+int update_file_info (FileRep *myFile) {
+FileInfo *myInfo = &(myFile->file_info);
+struct stat fStat;
+u_int32_t mySig;
+u_int8_t vers=0;
+
+/*
+  N.B.:  File opening and locking happens outside of this func!
+  The file must be opened read-write, and this func must have exclusive access to it.
+*/
+	lseek(myFile->fd_info, (off_t)0, SEEK_SET);
+	if (fstat (myFile->fd_info , &fStat) != 0) return (-2);
+	if (fStat.st_size == 276) vers = 1;
+
+	if (vers != 1) {
+		if ( read (myFile->fd_info,(void *)&(mySig),sizeof(mySig)) != sizeof(mySig) ) {
+			return (-2);
+		}
+		if (mySig != OME_IS_FILE_SIG) return (-3);
+
+		if ( read (myFile->fd_info,(void *)&(vers),sizeof(vers)) != sizeof(vers) ) {
+			return (-3);
+		}
+	}
+
+	lseek(myFile->fd_info, (off_t)0, SEEK_SET);
+	switch(vers){
+		case 1:
+			if ( read (myFile->fd_info,(void *)&(myInfo->sha1),sizeof(myInfo->sha1)) != sizeof(myInfo->sha1) ) 
+				return (0);
+			if ( read (myFile->fd_info,(void *)&(myInfo->name),sizeof(myInfo->name)) != sizeof(myInfo->name) )
+				return (0);
+			myInfo->mySig    = OME_IS_FILE_SIG;
+			myInfo->vers     = OME_IS_FILE_VER;
+			myInfo->ID       = myFile->ID;
+			myInfo->isAlias  = 0;
+			myInfo->nAliases = 0;
+		break;
+		
+		case OME_IS_FILE_VER:
+			if ( read (myFile->fd_info,(void *)myInfo,sizeof(FileInfo)) != sizeof(FileInfo) )
+				return (0);
+		break;
+	}
+	
+	lseek(myFile->fd_info, (off_t)0, SEEK_SET);
+	if ( write (myFile->fd_info,(void *)myInfo,sizeof(FileInfo)) != sizeof(FileInfo) ) {
+		return (0);
+	}
+	
+	return (1);
+}
+
+
 int GetFileInfo (FileRep *myFile) {
+u_int32_t mySig;
+u_int8_t vers=0;
+struct stat fStat;
+
+	/* Do nothing if version and signature are set */
+	if ( myFile->file_info.mySig == OME_IS_FILE_SIG && myFile->file_info.vers == OME_IS_FILE_VER)
+		return (1);
 
 	if ( (myFile->fd_info = open (myFile->path_info, O_RDONLY, 0600)) < 0) {
+		OMEIS_DoError ("Error opening FileID=%llu: %s.",
+			(unsigned long long)myFile->ID,strerror (errno));
 		return (-1);
 	}
+	/* We'll block here until others are finish writing */
+	lockRepFile (myFile->fd_rep,'r',0LL,0LL);
 
-	if ( read (myFile->fd_info,(void *)&(myFile->file_info),sizeof(FileInfo)) != sizeof(FileInfo) ) {
-		return (-2);
+	if (fstat (myFile->fd_info , &fStat) == -1) return (-2);
+	if (fStat.st_size == 276) vers = 1;
+	
+	if (vers != 1) {
+		if ( read (myFile->fd_info,(void *)&(mySig),sizeof(mySig)) != sizeof(mySig) ) {
+			OMEIS_DoError ("Error reading FileID=%llu: %s.",
+				(unsigned long long)myFile->ID,strerror (errno));
+			return (-2);
+		}
+		if (mySig != OME_IS_FILE_SIG) {
+			OMEIS_DoError ("Error reading FileID=%llu.  Bad signature : %s.",
+				(unsigned long long)myFile->ID,strerror (errno));
+			return (-3);
+		}
+
+		if ( read (myFile->fd_info,(void *)&(vers),sizeof(vers)) != sizeof(vers) ) {
+			OMEIS_DoError ("Error reading version for FileID=%llu: %s.",
+				(unsigned long long)myFile->ID,strerror (errno));
+			return (-4);
+		}
+	}
+	
+	lseek(myFile->fd_info, (off_t)0, SEEK_SET);
+	if (vers == 0) {
+		OMEIS_DoError ("Error reading version for FileID=%llu: %s.",
+			(unsigned long long)myFile->ID,strerror (errno));
+		return (-5);
 	}
 
-	close (myFile->fd_info);
-	myFile->fd_info = -1;
+	if (vers == OME_IS_FILE_VER) {
+	if ( read (myFile->fd_info,(void *)&(myFile->file_info),sizeof(FileInfo)) != sizeof(FileInfo) ) {
+			OMEIS_DoError ("Error reading info for FileID=%llu: %s.",
+				(unsigned long long)myFile->ID,strerror (errno));
+			return (-6);
+		}
+	} else {
+		/* re-open the file for exclusive read-write */
+		/* closing releases all locks */
+		close (myFile->fd_info);
+		myFile->fd_info = -1;
+		chmod (myFile->path_info,0600);	
+		if ( (myFile->fd_info = open (myFile->path_info, O_RDWR, 0600)) < 0) {
+			OMEIS_DoError ("Error opening info for FileID=%llu: %s.",
+				(unsigned long long)myFile->ID,strerror (errno));
+			return (-7);
+		}
+		/* We'll block here until we can write exclusively */
+		lockRepFile (myFile->fd_rep,'w',0LL,0LL);
+		/*
+		  At this point, another process could have already fixed this file,
+		  So update_file_info will determine the version again on its own.
+		*/
+		if (! update_file_info (myFile)) {
+			OMEIS_DoError ("Error updating info version for FileID=%llu: %s.",
+				(unsigned long long)myFile->ID,strerror (errno));
+			return (-8);
+		}
+		close (myFile->fd_info);
+		chmod (myFile->path_info,0400);
+		/* And release the lock also */
+		myFile->fd_info = -1;
+	}
+
+	if (myFile->fd_info > -1) {
+		close (myFile->fd_info);
+		myFile->fd_info = -1;
+	}
 
 	return (1);
+}
+
+int GetFileAliases (FileRep *myFile) {
+unsigned int i, nAliases;
+FileAlias *aliases,*myAlias;
+
+	/* return immediately if aliases is not NULL */
+	if (myFile->aliases) return (1);
+
+	/* Make sure the info struct is filled up */
+	GetFileInfo (myFile);
+
+	/* No aliases, so return 1 */
+	if ( (nAliases = myFile->file_info.nAliases) == 0 ) {
+		return (1);
+	}
+	
+	/* Get enough memory for all of the aliases */
+	if ( ! (myAlias = aliases = (FileAlias *)malloc (nAliases * sizeof (FileAlias))) ) {
+		close (myFile->fd_info);
+		myFile->fd_info = -1;
+		return (0);
+	}
+
+	/* Open the info file if necessary */
+	if (myFile->fd_info < 0) {
+		if ( (myFile->fd_info = open (myFile->path_info, O_RDONLY, 0600)) < 0) {
+			free (aliases);
+			close (myFile->fd_info);
+			myFile->fd_info = -1;
+			return (0);
+		}
+	}
+	
+	lseek(myFile->fd_info, (off_t)sizeof(FileInfo), SEEK_SET);
+	for (i = 0; i < nAliases; i++) {
+		if ( read (myFile->fd_info,(void *)myAlias,sizeof(FileAlias)) != sizeof(FileAlias) ) {
+			free (aliases);
+			close (myFile->fd_info);
+			myFile->fd_info = -1;
+			return (0);
+		}
+		myAlias++;
+	}
+	close (myFile->fd_info);
+	myFile->fd_info = -1;
+	myFile->aliases = aliases;
+	return (1);	
+	
+}
+
+
+OID check_aliases (OID ID, const char *name) {
+FileRep *myFile;
+unsigned int i, nAliases;
+OID found=0;
+FileAlias *alias;
+
+	/* Get a FileRep struct */
+	if (! (myFile = newFileRep (ID))) {
+		OMEIS_DoError ("Could not get a File object (out of memory?).");
+		return (0);
+	}
+
+	/* Read the original's info file */
+	GetFileInfo (myFile);
+
+	/* Check the name */
+	if (! strncmp (myFile->file_info.name,name,OMEIS_PATH_SIZE) ) {
+		freeFileRep (myFile);
+		return (ID);
+	}
+
+	/* No aliases, and name didn't match, so return 0 */
+	if ( (nAliases = myFile->file_info.nAliases) == 0 ) {
+		freeFileRep (myFile);
+		return (0);
+	}
+
+	GetFileAliases (myFile);
+	
+	if (myFile->aliases) {
+		alias = myFile->aliases;
+		for (i = 0; i < nAliases && !found; i++) {
+			if (! strncmp (alias->name,name,OMEIS_PATH_SIZE) )
+				found = alias->ID;
+			alias++;
+		}
+	}
+
+	freeFileRep (myFile);
+	return (found);
+
+
+}
+
+/* returns true on success */
+int make_alias (OID ID,FileRep *aliasFile) {
+FileRep *myFile;
+unsigned int i, nAliases;
+FileAlias alias;
+int myFlags;
+char backupName[OMEIS_PATH_SIZE];
+char relPath[OMEIS_PATH_SIZE];
+
+	/* Get a FileRep struct */
+	if (! (myFile = newFileRep (ID))) {
+		OMEIS_DoError ("Could not get a File object (out of memory?).");
+		return (0);
+	}
+
+	/* Get the both file's infos */
+	GetFileInfo (myFile);
+	GetFileInfo (aliasFile);
+
+	/* Close the info files if they're open */
+	if (myFile->fd_info >= 0) close (myFile->fd_info);
+	if (aliasFile->fd_info >= 0) close (aliasFile->fd_info);
+	myFile->fd_info = -1;
+	aliasFile->fd_info = -1;
+
+	/* chmod and reopen the both file's info with exclusive write access */
+	chmod (myFile->path_info,0600);	
+	if ( (myFile->fd_info = open (myFile->path_info, O_RDWR, 0600)) < 0) {
+		return (0);
+	}
+	lockRepFile (myFile->fd_rep,'w',0LL,0LL);
+	chmod (aliasFile->path_info,0600);	
+	if ( (aliasFile->fd_info = open (aliasFile->path_info, O_RDWR, 0600)) < 0) {
+		return (0);
+	}
+	lockRepFile (aliasFile->fd_rep,'w',0LL,0LL);
+
+	/* Copy the alias file's path, and rename it to a temporary backup (append ~) */
+	strncpy (backupName,aliasFile->path_rep,OMEIS_PATH_SIZE-1);
+	strcat (backupName,"~");
+	if (rename (aliasFile->path_rep,backupName) == -1) {
+		OMEIS_DoError ("Error making backup of FileID=%llu: %s.",
+			(unsigned long long)aliasFile->ID,strerror (errno));
+		return (0);
+	}
+
+	/* make a symlink between the original and the alias */
+	if (symlink (get_rel_path(myFile->path_rep,aliasFile->path_rep,relPath), aliasFile->path_rep) == -1) {
+		OMEIS_DoError ("Error making symlink from %s to %s: %s.",
+			aliasFile->path_rep,relPath,strerror (errno));
+		/* restore backup on error */
+		if (rename (backupName, aliasFile->path_rep) == -1) {
+			OMEIS_DoError ("Error restoring backup %s.",backupName);
+			return (0);
+		}
+		return (0);
+	}
+
+	/* Ditch the backup */
+	unlink (backupName);
+	
+	/* increment nAliases in the original, and set isAlias in the alias */
+	myFile->file_info.nAliases++;
+	aliasFile->file_info.isAlias = myFile->ID;
+
+	/* write both info structures back */
+	if ( write (myFile->fd_info,(void *)&(myFile->file_info),sizeof(FileInfo)) != sizeof(FileInfo) ) {
+		OMEIS_DoError ("Error writing info for original FileID=%llu: %s",
+			(unsigned long long)myFile->ID, strerror (errno));
+		return (0);
+	}
+	if ( write (aliasFile->fd_info,(void *)&(aliasFile->file_info),sizeof(FileInfo)) != sizeof(FileInfo) ) {
+		OMEIS_DoError ("Error writing info for alias FileID=%llu: %s",
+			(unsigned long long)aliasFile->ID, strerror (errno));
+		return (0);
+	}
+
+	/* Set write mode to append for the original */
+	if ( (myFlags = fcntl (myFile->fd_info, F_GETFL, 0)) < 0) {
+		OMEIS_DoError ("Error getting file flags for %s: %s",
+			myFile->path_info, strerror (errno));
+		return (0);
+	}
+	myFlags |= O_APPEND;
+	if ( fcntl (myFile->fd_info, F_SETFL, myFlags) < 0) {
+		OMEIS_DoError ("Error setting append file flag for %s: %s",
+			myFile->path_info, strerror (errno));
+		return (0);
+	}
+
+	/* set the FileAlias struct */
+	alias.ID = aliasFile->ID;
+	strncpy (alias.name,aliasFile->file_info.name,OMEIS_PATH_SIZE);
+	
+	/* append the alias struct to the file */
+	if ( write (myFile->fd_info,(void *)&(alias),sizeof(alias)) != sizeof(alias) ) {
+		OMEIS_DoError ("Error writing alias info for original FileID=%llu: %s",
+			(unsigned long long)myFile->ID, strerror (errno));
+		return (0);
+	}
+	
+	/* Close both file's infos and chmod back to read-only */
+	chmod (myFile->path_info,0400);
+	close (myFile->fd_info);
+	myFile->fd_info = -1;
+	chmod (aliasFile->path_info,0400);
+	close (aliasFile->fd_info);
+	aliasFile->fd_info = -1;
+
+	/* return true */
+	return (1);
+
+}
+
+
+char *get_rel_path (const char *toPath, const char *fromPath, char *pathBuf) {
+char *tmp1,*tmp2;
+int nBack=0;
+
+	strcpy (pathBuf,"");
+	tmp1 = strrchr (toPath,'/');
+	if (tmp1) tmp1++;
+	while (*toPath && *fromPath && *toPath == *fromPath && toPath != tmp1) {
+		toPath++;
+		fromPath++;
+	}
+
+	while (*fromPath) {
+		if (*fromPath++ == '/') nBack++;
+	}
+	tmp2 = pathBuf + (nBack*3);
+
+	while (nBack) {
+		strcat (pathBuf,"../");
+		nBack--;
+	}
+	
+	while (*toPath) {
+		*tmp2++ = *toPath++;
+	}
+	*tmp2 = '\0';
+	
+	return (pathBuf);
 }
 
 /*
   Call this only to balance a call to NewFile()
 */
 OID FinishFile (FileRep *myFile) {
-OID existOID;
+OID existOID, aliasOID;
 
 	/* Get SHA1 */
 	if ( get_md_from_buffer (myFile->file_buf, myFile->size_rep, myFile->file_info.sha1) < 0 ) {
@@ -313,17 +706,39 @@ OID existOID;
 	if ( (existOID = sha1DB_get (myFile->DB, myFile->file_info.sha1)) ) {
 		sha1DB_close (myFile->DB);
 		myFile->DB = NULL;
-		DeleteFile (myFile);
-		myFile->ID = existOID;
-		return (existOID);
+		/* Check the original for duplicate alias */
+		if ( (aliasOID = check_aliases (existOID,myFile->file_info.name)) ) {
+			DeleteFile (myFile);
+			myFile->ID = aliasOID;
+			myFile->file_info.ID = aliasOID;
+			return (aliasOID);
+		} else {
+			/* unlinks myFile, makes a symlink to original, records alias in original,
+			  file_info.isAlias set to existOID. un-mmapps the file.
+			  myFile->file_info is not written.
+			  if anything goes wrong here, this function will return 0.
+			  If everything's OK, it will return 1.
+			*/
+			if (! (make_alias (existOID,myFile)) ) {
+				DeleteFile (myFile);
+				return (0);
+			}			
+		}
 	}
 
+	chmod (myFile->path_info,0600);
 	if ( (myFile->fd_info = open (myFile->path_info, O_RDWR, 0600)) < 0) {
 		DeleteFile (myFile);
+		sha1DB_close (myFile->DB);
+		myFile->DB = NULL;
 		return (0);
 	}
 
 	if ( write (myFile->fd_info,(void *)&(myFile->file_info),sizeof(FileInfo)) != sizeof(FileInfo) ) {
+		OMEIS_DoError ("In FinishFile, Error writing info for FileID=%llu: %s",
+			(unsigned long long)myFile->ID, strerror (errno) );
+		sha1DB_close (myFile->DB);
+		myFile->DB = NULL;
 		DeleteFile (myFile);
 		return (0);
 	}
@@ -333,6 +748,10 @@ OID existOID;
 
 	if (myFile->is_mmapped) {
 		if (msync (myFile->file_buf , myFile->size_rep , MS_SYNC) != 0) {
+			OMEIS_DoError ("In FinishFile, Error msynching FileID=%lly: %s",
+				(unsigned long long)myFile->ID, strerror (errno) );
+			sha1DB_close (myFile->DB);
+			myFile->DB = NULL;
 			DeleteFile (myFile);
 			return (0);
 		}
@@ -348,16 +767,20 @@ OID existOID;
 	}
 
 	/* put the SHA1 in the DB */
-	if ( sha1DB_put (myFile->DB, myFile->file_info.sha1, myFile->ID) ) {
+	if (! existOID) {
+		if ( sha1DB_put (myFile->DB, myFile->file_info.sha1, myFile->ID) ) {
+			OMEIS_DoError ("In FinishFile, Error writing SHA1-DB for FileID=%llu: %s",
+				(unsigned long long)myFile->ID, strerror (errno) );
+			sha1DB_close (myFile->DB);
+			myFile->DB = NULL;
+			DeleteFile (myFile);
+			return (0);
+		}
+		/* Close the DB (and release the exclusive lock) */
 		sha1DB_close (myFile->DB);
 		myFile->DB = NULL;
-		DeleteFile (myFile);
-		return (0);
 	}
 
-	/* Close the DB (and release the exclusive lock) */
-	sha1DB_close (myFile->DB);
-	myFile->DB = NULL;
 
 
 	chmod (myFile->path_info,0400);
@@ -390,7 +813,7 @@ FILE *infile;
     if (infile) {
         nIO = fread (myFile->file_buf,1,size,infile);
         if (nIO != size) {
-            fprintf (stderr,"Couldn't finish writing uploaded file %s (ID=%llu).  Wrote %lu, expected %lu\n",
+        	OMEIS_DoError ("Couldn't finish writing uploaded file %s (ID=%llu).  Wrote %lu, expected %lu",
                      filename,(unsigned long long)ID,(unsigned long)nIO,(unsigned long)size);
             DeleteFile (myFile);
 			freeFileRep (myFile);
